refactor: refactor user form
