+++ src/components/AddUserForm/AddUserForm.tsx
@@ -4,12 +4,7 @@ import styles from './styles.module.scss';
 import { useAppDispatch } from '../../hooks/redux';
 import { createUser } from '../../redux/asyncActions/users';
 import { Access, Modules, UserDto } from '../../models/User/UserDto';
-import validation from '../../constants/user-validation';
-import usePasswordVisibility from '../../hooks/usePasswordVisibility';
-import SectionCheckboxes from '../SectionCheckboxes';
-import Input from '../../UI/Input';
-import Eye from '../../UI/Eye';
-import Button from '../../UI/Button';
+import UserForm from '../UserForm';
 
 interface Props {
   close: () => void;
@@ -17,8 +12,6 @@ interface Props {
 
 const AddUserForm = ({ close }: Props) => {
   const dispatch = useAppDispatch();
-  const { username, password, email } = validation;
-  const { isVisible, toggle } = usePasswordVisibility();
 
   const defaultValues = {
     access: Object.values(Modules).reduce((acc, cur) => {
@@ -29,13 +22,12 @@ const AddUserForm = ({ close }: Props) => {
 
   const {
     register,
-    control,
     handleSubmit,
     reset,
     formState: { errors },
   } = useForm<UserDto>({
     mode: 'all',
-    defaultValues ,
+    defaultValues,
   });
 
   const onSubmit: SubmitHandler<UserDto> = (data: UserDto) => {
@@ -44,58 +36,9 @@ const AddUserForm = ({ close }: Props) => {
     reset();
   };
 
-  const sectionPermissions = Object.values(Modules).map((section, index) => (
-    <SectionCheckboxes
-      key={index}
-      section={section}
-      name={`access.${section}`}
-      register={register}
-    />
-  ));
-
   return (
     <div className={styles.formWrapper}>
-      <form className={styles.form}>
-        <Input
-          name={'username'}
-          label={'Username'}
-          register={register}
-          options={{
-            required: username.required,
-            minLength: { value: 3, message: username.message },
-          }}
-          errors={errors}
-        />
-        <Input
-          name={'email'}
-          label={'Email'}
-          register={register}
-          options={{
-            required: email.required,
-            pattern: { value: email.pattern, message: email.message },
-          }}
-          errors={errors}
-        />
-        <Input
-          name={'password'}
-          label={'Password'}
-          register={register}
-          options={{
-            required: password.required,
-            minLength: { value: 4, message: password.message },
-          }}
-          errors={errors}
-          isVisible={isVisible}
-        >
-          <Eye handleClick={toggle} />
-        </Input>
-        {sectionPermissions}
-        <div onClick={handleSubmit(onSubmit)} className={styles.button}>
-          <Button inverted={false} padding={'20px 30px'}>
-            Save
-          </Button>
-        </div>
-      </form>
+      <UserForm onSubmit={handleSubmit(onSubmit)} errors={errors} register={register} />
     </div>
   );
 };
