+++ src/components/AddUserForm/styles.module.scss
@@ -13,16 +13,3 @@
     padding: 4.5em 3.5em;
   }
 }
-
-.form {
-  width: 100%;
-}
-
-.button {
-  width: 100%;
-}
-
-.roles {
-  display: flex;
-  gap: 10px;
-}
