+++ src/components/EditUserForm/EditUserForm.tsx
@@ -4,14 +4,9 @@ import styles from './styles.module.scss';
 import icons from '../../assets/icons.svg';
 import { useAppDispatch } from '../../hooks/redux';
 import { updateUser } from '../../redux/asyncActions/users';
-import { Modules, UserDto } from '../../models/User/UserDto';
+import { UserDto } from '../../models/User/UserDto';
 import { User } from '../../models/User/User';
-import Input from '../../UI/Input';
-import Eye from '../../UI/Eye';
-import Button from '../../UI/Button';
-import validation from '../../constants/user-validation';
-import usePasswordVisibility from '../../hooks/usePasswordVisibility';
-import SectionCheckboxes from '../SectionCheckboxes';
+import UserForm from '../UserForm';
 
 interface Props {
   close: () => void;
@@ -20,9 +15,6 @@ interface Props {
 
 const EditUserForm = ({ close, user }: Props) => {
   const dispatch = useAppDispatch();
-  const { username, password, email } = validation;
-  const { isVisible, toggle } = usePasswordVisibility();
-
   const { _id, ...defaultValues } = user;
 
   const {
@@ -41,58 +33,9 @@ const EditUserForm = ({ close, user }: Props) => {
     reset();
   };
 
-  const sectionPermissions = Object.values(Modules).map((section, index) => (
-    <SectionCheckboxes
-      key={index}
-      section={section}
-      name={`access.${section}`}
-      register={register}
-    />
-  ));
-
   return (
     <div className={styles.formWrapper}>
-      <form className={styles.form}>
-        <Input
-          name={'username'}
-          label={'Username'}
-          register={register}
-          options={{
-            required: username.required,
-            minLength: { value: 3, message: username.message },
-          }}
-          errors={errors}
-        />
-        <Input
-          name={'email'}
-          label={'Email'}
-          register={register}
-          options={{
-            required: email.required,
-            pattern: { value: email.pattern, message: email.message },
-          }}
-          errors={errors}
-        />
-        <Input
-          name={'password'}
-          label={'Password'}
-          register={register}
-          options={{
-            required: password.required,
-            minLength: { value: 4, message: password.message },
-          }}
-          errors={errors}
-          isVisible={isVisible}
-        >
-          <Eye handleClick={toggle} />
-        </Input>
-        {sectionPermissions}
-        <div onClick={handleSubmit(onSubmit)} className={styles.button}>
-          <Button inverted={false} padding={'20px 30px'}>
-            Save
-          </Button>
-        </div>
-      </form>
+      <UserForm onSubmit={handleSubmit(onSubmit)} errors={errors} register={register} />
       <svg className={styles.closeButton} onClick={close}>
         <use xlinkHref={`${icons}#cross`} />
       </svg>
